+++ src/main/java/tdd/tddproject/rin/item/repository/ItemRepository.java
@@ -1,4 +1,4 @@
-package tdd.tddproject.rin.repository;
+package tdd.tddproject.rin.item.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
