+++ src/main/java/tdd/tddproject/rin/mapper/ItemMapper.java
@@ -3,7 +3,7 @@ package tdd.tddproject.rin.mapper;
 import org.mapstruct.Mapper;
 import tdd.tddproject.domain.entity.item.Item;
 import tdd.tddproject.global.mapper.GenericMapper;
-import tdd.tddproject.rin.dto.ItemDto;
+import tdd.tddproject.rin.item.dto.ItemDto;
 
 @Mapper(componentModel = "spring")
 public interface ItemMapper extends GenericMapper<ItemDto, Item> {
